+++ src/config/__init__.py
@@ -1,0 +1,7 @@
+"""
+Configuration management modules for ClipperTool
+"""
+
+from .config_manager import ConfigManager
+
+__all__ = ['ConfigManager']
