+++ src/core/__init__.py
@@ -1,0 +1,7 @@
+"""
+Core functionality modules for ClipperTool
+"""
+
+from .clipboard_processor import ClipboardProcessor
+
+__all__ = ['ClipboardProcessor']
