+++ src/ui/__init__.py
@@ -1,0 +1,7 @@
+"""
+User interface modules for ClipperTool
+"""
+
+from src.ui.main_window import ClipperToolUI
+
+__all__ = ['ClipperToolUI']
