+++ src/ui/constants.py
@@ -1,20 +1,39 @@
-import os
+"""
+UI Constants and styling definitions
+"""
 
+# Modern high-tech color scheme
 COLORS = {
-    'bg_primary': '#1e1e2f',
-    'bg_secondary': '#2e2e3f',
-    'bg_card': '#2a2a3a',
-    'text_primary': '#ffffff',
-    'text_secondary': '#aaaaaa',
+    'bg_primary': '#0f0f0f',
+    'bg_secondary': '#1a1a1a',
+    'bg_card': '#252525',
     'accent_gold': '#d4af37',
-    'accent_blue': '#3a8ee6',
-    'accent_red': '#ff5c57',
-    'success': '#28a745',
-    'danger': '#dc3545',
-    'accent_green': '#4caf50',
+    'accent_gold_hover': '#f4cf47',
+    'accent_blue': '#00d4ff',
+    'accent_green': '#00ff88',
+    'accent_red': '#ff4757',
+    'text_primary': '#ffffff',
+    'text_secondary': '#b0b0b0',
+    'border': '#333333',
+    'success': '#00ff88',
+    'danger': '#ff4757',
+    'warning': '#ffa726'
 }
 
-FILTERS_FOLDER = os.path.join(os.path.expanduser("~"), ".clippertool_filters")
+# Window configuration
+WINDOW_CONFIG = {
+    'geometry': '800x650',
+    'min_width': 700,
+    'min_height': 600
+}
 
-# Make sure the filters folder exists
-os.makedirs(FILTERS_FOLDER, exist_ok=True)
+# Font configuration
+FONTS = {
+    'title': ('Consolas', 24, 'bold'),
+    'header': ('Consolas', 12, 'bold'),
+    'body': ('Consolas', 11),
+    'button': ('Consolas', 10),
+    'button_large': ('Consolas', 14, 'bold'),
+    'button_main': ('Consolas', 16, 'bold'),
+    'footer': ('Consolas', 9)
+}
