+++ src/utils/__init__.py
@@ -1,0 +1,7 @@
+"""
+Utility modules for ClipperTool
+"""
+
+from .resources import resource_path
+
+__all__ = ['resource_path']
